# Optimize git resource for fixed revision