`git fetch origin` can be very slow. For example,

```bash
$ git remote -v
origin  github.com:rbenv/rbenv (fetch)
origin  github.com:rbenv/rbenv (push)
$ time git fetch origin
git fetch origin  0.02s user 0.02s system 0% cpu 3.285 total
```

So I want to skip `git fetch origin` if HEAD is the same as specified revision.
With this change, following recipe execution time is optimized from 4.14s to 1.20s.

```rb
git "/Users/takashi-kokubun/.rbenv" do
  repository "git://github.com/rbenv/rbenv.git"
  revision "20755cfc153dd0bf802524cbb313f834a59c0f30"
end
```